+++ scala/core/src/main/scala/com/thenewmotion/mobilityid/checkDigit.scala
@@ -21,8 +21,8 @@ import scala.annotation.tailrec
 
 private[mobilityid] object CheckDigitIso {
 
-  import MatrixUtil._
   import LookupTables._
+  import MatrixUtil._
 
   def apply(code: String): Char = applyToUpperCaseString(code.toUpperCase)
 
@@ -32,13 +32,12 @@ private[mobilityid] object CheckDigitIso {
     require(code.forall(_.isAsciiUpperOrDigit), "Code must consist of uppercase ASCII letters and digits")
 
     @annotation.tailrec
-    def sumEq(ps: Array[Matrix], f: Matrix => Vec, v: Vec = Vec(0, 0), i: Int = 0): Vec = {
+    def sumEq(ps: Array[Matrix], f: Matrix => Vec, v: Vec = Vec(0, 0), i: Int = 0): Vec =
       if (i < ps.length) {
         val ch = code.charAt(i)
         val qr = f(encoding.getOrElse(ch, sys.error(s"Invalid character: $ch.")))
         sumEq(ps, f, v + (qr * ps(i)), i + 1)
       } else v
-    }
 
     val t1 = sumEq(p1s, m => Vec(m.m11, m.m12))
     val t2 = sumEq(p2s, m => Vec(m.m21, m.m22)) * negP2minus15
@@ -55,18 +54,42 @@ private[mobilityid] object LookupTables {
   def encode(mx: Matrix): Int = mx.m11 + (mx.m12 << 1) + (mx.m21 << 2) + (mx.m22 << 4)
 
   val encoding: Map[Char, Matrix] = Map[Char, Int](
-    '0' -> 0, '1' -> 16, '2' -> 32,
-    '3' -> 4, '4' -> 20, '5' -> 36,
-    '6' -> 8, '7' -> 24, '8' -> 40,
-    '9' -> 2, 'A' -> 18, 'B' -> 34,
-    'C' -> 6, 'D' -> 22, 'E' -> 38,
-    'F' -> 10, 'G' -> 26, 'H' -> 42,
-    'I' -> 1, 'J' -> 17, 'K' -> 33,
-    'L' -> 5, 'M' -> 21, 'N' -> 37,
-    'O' -> 9, 'P' -> 25, 'Q' -> 41,
-    'R' -> 3, 'S' -> 19, 'T' -> 35,
-    'U' -> 7, 'V' -> 23, 'W' -> 39,
-    'X' -> 11, 'Y' -> 27, 'Z' -> 43
+    '0' -> 0,
+    '1' -> 16,
+    '2' -> 32,
+    '3' -> 4,
+    '4' -> 20,
+    '5' -> 36,
+    '6' -> 8,
+    '7' -> 24,
+    '8' -> 40,
+    '9' -> 2,
+    'A' -> 18,
+    'B' -> 34,
+    'C' -> 6,
+    'D' -> 22,
+    'E' -> 38,
+    'F' -> 10,
+    'G' -> 26,
+    'H' -> 42,
+    'I' -> 1,
+    'J' -> 17,
+    'K' -> 33,
+    'L' -> 5,
+    'M' -> 21,
+    'N' -> 37,
+    'O' -> 9,
+    'P' -> 25,
+    'Q' -> 41,
+    'R' -> 3,
+    'S' -> 19,
+    'T' -> 35,
+    'U' -> 7,
+    'V' -> 23,
+    'W' -> 39,
+    'X' -> 11,
+    'Y' -> 27,
+    'Z' -> 43
   ).map { case (k, v) => k -> decode(v) }
 
   val decoding: Map[Matrix, Char] = encoding.map(_.swap)
@@ -76,8 +99,12 @@ private[mobilityid] object MatrixUtil {
 
   case class Matrix(m11: Int, m12: Int, m21: Int, m22: Int) {
     def *(m: Matrix): Matrix =
-      Matrix(m11 * m.m11 + m12 * m.m21, m11 * m.m12 + m12 * m.m22,
-        m21 * m.m11 + m22 * m.m21, m21 * m.m12 + m22 * m.m22)
+      Matrix(
+        m11 * m.m11 + m12 * m.m21,
+        m11 * m.m12 + m12 * m.m22,
+        m21 * m.m11 + m22 * m.m21,
+        m21 * m.m12 + m22 * m.m22
+      )
   }
 
   case class Vec(v1: Int, v2: Int) {
@@ -96,8 +123,7 @@ private[mobilityid] object MatrixUtil {
 
 private[mobilityid] object CheckDigitDin {
 
-  private val toNumericValue =
-    (('0' to '9') ++ ('A' to 'Z')).zipWithIndex.toMap
+  private val toNumericValue = (('0' to '9') ++ ('A' to 'Z')).zipWithIndex.toMap
 
   def apply(contractId: String): Char = {
     def mult(value: Int, coeff: Int) = value * math.pow(2, coeff).toInt
@@ -106,7 +132,7 @@ private[mobilityid] object CheckDigitDin {
     val lookupResults = theString map toNumericValue
 
     @tailrec
-    def go(rest: IndexedSeq[Int], acc: Int, coefficient: Int): Int = {
+    def go(rest: IndexedSeq[Int], acc: Int, coefficient: Int): Int =
       if (rest.isEmpty) acc
       else {
         val current = rest.head
@@ -121,12 +147,9 @@ private[mobilityid] object CheckDigitDin {
 
         go(rest.tail, acc + stepResult, newCoefficient)
       }
-    }
 
     val sum = go(lookupResults, 0, 0)
     val mod = sum % 11
     if (mod >= 10) 'X' else Character.forDigit(mod, 10)
   }
 }
-
-