+++ scala/core/src/main/scala/com/thenewmotion/mobilityid/operatorId.scala
@@ -52,5 +52,6 @@ object OperatorIdDin {
     if (isValid(id)) {
       OperatorIdDinImpl(id.toUpperCase)
     } else throw new IllegalArgumentException(
-        "OperatorId must have a length of 3-6 chars and be digits")
+      "OperatorId must have a length of 3-6 chars and be digits"
+    )
 }